requirements.txt
pytorch
timm
torchvision
skimage
opencv-python